--- a/be/src/olap/rowset/rowset_writer.h
+++ b/be/src/olap/rowset/rowset_writer.h
@@ -20,6 +20,8 @@
 #include <gen_cpp/olap_file.pb.h>
 #include <gen_cpp/types.pb.h>
 
+#include <optional>
+
 #include "common/factory_creator.h"
 #include "gutil/macros.h"
 #include "olap/column_mapping.h"
@@ -37,6 +39,7 @@ struct FlushContext {
     ENABLE_FACTORY_CREATOR(FlushContext);
     TabletSchemaSPtr flush_schema = nullptr;
     const vectorized::Block* block = nullptr;
+    std::optional<int32_t> seq = std::nullopt;
 };
 
 class RowsetWriter {
@@ -98,6 +101,8 @@ public:
     }
 
     virtual int32_t get_atomic_num_segment() const = 0;
+
+    virtual int32_t allocate_seq_id() = 0;
 
     virtual bool is_doing_segcompaction() const = 0;