## Proposed changes

Assign a `flush_serial_id` to `MemTable` early in `DeltaWriter`, and exec the flush in parallel.

In this way, we can improve the stream load peformance of 1 tablet unique key tables by around 100%.

Stream load 7.5G tpch lineitem table into a 1 tablet unique key table (3 BE, 16C 64G):

| replicas | load time (before) | load time (after) |
| --- | --- | --- |
| 1 | 118s | 48s |
| 3 | 161s | 60s |

## Further comments

If this is a relatively large or complex change, kick off the discussion at [dev@doris.apache.org](mailto:dev@doris.apache.org) by explaining why you chose the solution you did and what alternatives you considered, etc...